Revert "169 feat(client): add UI to delete entry"
