+++ apps/client/package.json
@@ -12,7 +12,6 @@
   },
   "dependencies": {
     "@radix-ui/react-accordion": "^1.1.2",
-    "@radix-ui/react-alert-dialog": "^1.0.5",
     "@tiptap/extension-bubble-menu": "^2.2.4",
     "@tiptap/extension-document": "^2.2.4",
     "@tiptap/extension-floating-menu": "^2.4.0",
