--- apps/client/src/components/feature/delete-entry-modal/index.ts
@@ -1,1 +1,0 @@
-export * from "./delete-entry-modal.component";
