+++ apps/client/src/components/feature/entry-search-modal/entry-list/entry-list.component.tsx
@@ -42,12 +42,8 @@ export function EntryList() {
 
         if (error || !entryList) throw new Error("error getting entry");
 
-        const availabelEntries = entryList.filter(
-          (entry) => typeof entry.end_date !== "string",
-        );
-
         setEntries(
-          availabelEntries.sort((entry, nextEntry) =>
+          entryList.sort((entry, nextEntry) =>
             new Date(entry.updated_at) > new Date(nextEntry.updated_at)
               ? -1
               : 1,
@@ -88,21 +84,19 @@ export function EntryList() {
             }}
             value={entry.title}
           >
-            <article>
-              <Link
-                href={ClientRoutingService.app.entries.readById(entry.id)}
-                className="w-full flex items-center gap-2 text-base"
-                onClick={handleClickOnEntry}
-              >
-                <File className="mr-2 h-5 w-5" />
-                <div className="flex flex-col">
-                  <h1 className="text-sm">{entry.title}</h1>
-                  <p className="text-neutral-400 text-sm">
-                    {moment(entry.updated_at).startOf("seconds").fromNow()}
-                  </p>
-                </div>
-              </Link>
-            </article>
+            <Link
+              href={ClientRoutingService.app.entries.readById(entry.id)}
+              className="w-full flex items-center gap-2 text-base"
+              onClick={handleClickOnEntry}
+            >
+              <File className="mr-2 h-5 w-5" />
+              <div className="flex flex-col">
+                <span className="text-sm">{entry.title}</span>
+                <p className="text-neutral-400 text-sm">
+                  {moment(entry.updated_at).startOf("seconds").fromNow()}
+                </p>
+              </div>
+            </Link>
           </CommandItem>
         ))}
     </>
