+++ apps/client/src/components/ui/button/index.ts
@@ -1,2 +1,1 @@
 export * from "./button.component";
-export * from "./button.models";
