--- apps/client/src/components/ui/modal/index.ts
@@ -1,1 +1,0 @@
-export * from "./modal.component";
