+++ apps/client/src/models/routing/api/api-routing.model.ts
@@ -22,7 +22,6 @@ export class ApiRoutingService {
       entry: {
         createEntry: this.getPath("journal"),
         readEntryById: (entryId: string) => this.getPath(`journal/${entryId}`),
-        deleteById: (entryId: string) => this.getPath(`journal/${entryId}`),
         updateEntryById: (entryId: string) =>
           this.getPath(`journal/${entryId}`),
         getEntryList: this.getPath("journal/"),
